+++ packages/widgets/user-profile-widget/CHANGELOG.md
@@ -2,6 +2,51 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.15.3](https://github.com/descope/descope-js/compare/user-profile-widget-0.15.2...user-profile-widget-0.15.3) (2026-08-09)
+
+### Dependency Updates
+
+* `core-js-sdk` updated to version `2.67.1`
+* `web-js-sdk` updated to version `1.51.4`
+* `web-component` updated to version `4.0.1`
+* `sdk-mixins` updated to version `0.24.0`
+## [0.15.2](https://github.com/descope/descope-js/compare/user-profile-widget-0.15.1...user-profile-widget-0.15.2) (2026-08-03)
+
+### Dependency Updates
+
+* `web-component` updated to version `4.0.0`
+## [0.15.1](https://github.com/descope/descope-js/compare/user-profile-widget-0.15.0...user-profile-widget-0.15.1) (2026-08-03)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.70.7`
+## [0.15.0](https://github.com/descope/descope-js/compare/user-profile-widget-0.14.6...user-profile-widget-0.15.0) (2026-08-02)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.70.6`
+
+### Features
+
+* **widgets:** recovery email/phone in user management + user profile widgets ([#1445](https://github.com/descope/descope-js/issues/1445)) ([0ea68f8](https://github.com/descope/descope-js/commit/0ea68f875c69ba9a391ff3af282ad1b0d2336256)), closes [descope/backend#1894](https://github.com/descope/backend/issues/1894) [descope/console-app#5598](https://github.com/descope/console-app/issues/5598) [descope/content#2136](https://github.com/descope/content/issues/2136) [descope/etc#15631](https://github.com/descope/etc/issues/15631)
+
+## [0.14.6](https://github.com/descope/descope-js/compare/user-profile-widget-0.14.5...user-profile-widget-0.14.6) (2026-07-26)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.70.5`
+* `sdk-mixins` updated to version `0.23.2`
+* `sdk-component-drivers` updated to version `0.14.1`
+
+### Bug Fixes
+
+* **user-profile-widget:** rebuild preloaded modal flow on value change RELEASE ([#1447](https://github.com/descope/descope-js/issues/1447)) ([3bcebc1](https://github.com/descope/descope-js/commit/3bcebc1ea8255cf2f2940cdef48babcf3d04fb86)), closes [descope/etc#17202](https://github.com/descope/etc/issues/17202)
+
+## [0.14.5](https://github.com/descope/descope-js/compare/user-profile-widget-0.14.4...user-profile-widget-0.14.5) (2026-07-22)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.70.4`
 ## [0.14.4](https://github.com/descope/descope-js/compare/user-profile-widget-0.14.3...user-profile-widget-0.14.4) (2026-07-16)
 
 ### Dependency Updates